+++ typewriter-test/tests/compile_tests.rs
@@ -1,0 +1,5 @@
+#[test]
+fn compile_error_tests() {
+    let t = trybuild::TestCases::new();
+    t.compile_fail("tests/compile_error/*.rs");
+}